# cmd/compile: rewrite generic XOR -1 x into NOT x