… src/cmd/compile/internal/ssa/gen/generic.rules