This patch enable rewriting manual twos complement sign flips into NEG
instructions.

That yields a ~2X performance improvement on AMD64 and compacter code.
So this is a win on all tables.

This was not the case previously because even tho the
(ADD (NOT x) 1) => (NEG X) rule already existed.
The (XOR -1 x) => (NOT x) didn't.

Benchmark:
  func BenchmarkManualNeg(b *testing.B) {
    for i := uint64(b.N); i != 0; i-- {
      doNotOptimise = i^(1<<64-1) + 1
    }
  }

  func BenchmarkManualNot(b *testing.B) {
    for i := uint64(b.N); i != 0; i-- {
      doNotOptimise = i^(1<<64-1)
    }
  }

Before:
  goos: linux
  goarch: amd64
  cpu: AMD Ryzen 5 3600 6-Core Processor
  BenchmarkManualNeg-12	1000000000	0.4982 ns/op
  BenchmarkManualNot-12	1000000000	0.4952 ns/op # 2 cycles ?

After:
  BenchmarkManualNeg-12	1000000000	0.2481 ns/op
  BenchmarkManualNot-12	1000000000	0.2500 ns/op